scripts/analysis_bundle/requirements.txt
# Python 3.9+ recommended. Install: pip install -r requirements.txt
numpy>=1.20
pandas>=1.3
matplotlib>=3.5
seaborn>=0.12
biopython>=1.79
